feat: modify models
