--- Siksha/Domain/Models/DailyMenus.swift
@@ -1,16 +1,0 @@
-//
-//  DailyMenus.swift
-//  Siksha
-//
-//  Created by Jihyeon on 2/13/26.
-//
-
-import Foundation
-
-struct DailyMenus {
-    let date: String
-    let dateType: DateType
-    let breakfast: [RestaurantModel]
-    let lunch: [RestaurantModel]
-    let dinner: [RestaurantModel]
-}
