--- Siksha/Domain/Models/DateType.swift
@@ -1,14 +1,0 @@
-//
-//  DateType.swift
-//  Siksha
-//
-//  Created by Jihyeon on 2/13/26.
-//
-
-import Foundation
-
-enum DateType {
-    case weekday
-    case saturday
-    case holiday
-}
