+++ Siksha/Domain/Models/MenuModel.swift
@@ -10,10 +10,10 @@ import Foundation
 struct MenuModel: Identifiable {
     let id: Int
     let code: String
-    let nameKr: String?
-    let nameEn: String?
-    let price: Int?
-    let score: Double?
+    let nameKr: String
+    let nameEn: String
+    let price: Int
+    let score: Double
     let reviewCount: Int
     let isLiked: Bool
     let likeCount: Int
